finished part 1

+++ Matrix.cpp
@@ -11,7 +11,7 @@ Matrix::~Matrix() {}
 Matrix Matrix::operator+(const double &value) const {return Matrix(empty, 1, 1);}
 Matrix &Matrix::operator+=(const double &value) {return *this;}
 const Matrix Matrix::operator+() const {return *this;}
-Matrix operator+(const int &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
+Matrix operator+(const double &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
 
 // matrix addition operators (+)
 Matrix Matrix::operator+(const Matrix &mat) const {return Matrix(empty, 1, 1);}
@@ -21,7 +21,7 @@ Matrix &Matrix::operator+=(const Matrix &mat) {return *this;}
 Matrix Matrix::operator-(const double &value) const {return Matrix(empty, 1, 1);}
 Matrix &Matrix::operator-=(const double &value) {return *this;}
 const Matrix Matrix::operator-() const {return *this;}
-Matrix operator-(const int &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
+Matrix operator-(const double &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
 
 // matrix subtraction operators (-)
 Matrix Matrix::operator-(const Matrix &mat) const {return Matrix(empty, 1, 1);}
@@ -30,7 +30,7 @@ Matrix &Matrix::operator-=(const Matrix &mat) {return *this;}
 // real value multiplaction operators (*)
 Matrix Matrix::operator*(const double &value) const {return Matrix(empty, 1, 1);}
 Matrix &Matrix::operator*=(const double &value) {return *this;}
-Matrix operator*(const int &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
+Matrix operator*(const double &value, const Matrix &mat) {return Matrix(empty, 1, 1);}
 
 // matrix multiplaction operators (*)
 Matrix Matrix::operator*(const Matrix &mat) const {return Matrix(empty, 1, 1);}
